+++ Tests/NukeTests/ImagePipelineTests/ImagePipelineProgressiveDecodingTests.swift
@@ -186,7 +186,7 @@ struct ImagePipelineProgressiveDecodingTests {
         dataLoader.servesFirstChunkAutomatically = false
         var recordedPreviews: [ImageResponse] = []
         let task = pipeline.imageTask(with: request)
-        let stream = await task.subscribedPreviews()
+        let stream = task.previews
         dataLoader.resume()
         for try await preview in stream {
             #expect(preview.image.cgImage?.width == 45)
@@ -211,7 +211,7 @@ struct ImagePipelineProgressiveDecodingTests {
         dataLoader.servesFirstChunkAutomatically = false
         var recordedPreviews: [ImageResponse] = []
         let task = pipeline.imageTask(with: request)
-        let stream = await task.subscribedPreviews()
+        let stream = task.previews
         dataLoader.resume()
         for try await preview in stream {
             #expect(preview.image.nk_test_processorIDs.count == 1)
@@ -392,7 +392,7 @@ struct ImagePipelineProgressiveDecodingTests {
         // WHEN/THEN
         dataLoader.servesFirstChunkAutomatically = false
         let task = pipeline.imageTask(with: request)
-        let stream = await task.subscribedPreviews()
+        let stream = task.previews
         dataLoader.resume()
         var previewScale: CGFloat?
         for try await preview in stream {